refactor: move FetchAliases. delete KnomadicDsl
