--- src/main/kotlin/nz/adjmunro/outcome/KnomadicDsl.kt
@@ -1,12 +1,0 @@
-package nz.adjmunro.outcome
-
-@Target(
-    AnnotationTarget.FILE,
-    AnnotationTarget.TYPEALIAS,
-    AnnotationTarget.CLASS,
-    AnnotationTarget.CONSTRUCTOR,
-    AnnotationTarget.FUNCTION,
-    AnnotationTarget.PROPERTY,
-)
-@DslMarker
-public annotation class KnomadicDsl
